+++ Changes.md
@@ -4,9 +4,12 @@ This file is a manually maintained list of changes for each release. Feel free
 to add your changes here when sending pull requests. Also send corrections if
 you spot any mistakes.
 
-## Unreleased
-* Adding __dirname before the directory not longer necessary
-* Added name and path to resolve function
+## HEAD
+* Support for relative paths, `__dirname` no longer necessary #42
+* Pass name and filepath to resolve function
+
+## v2.1.0 (2016-12-09)
+* Accept a function for "filter" option #27 #31
 
 ## v2.0.0 (2015-10-17)